Refactor DTOs and remove obsolete classes
- Deleted CvuValidationResponseDTO as it is no longer needed.
- Introduced new DTOs: WithdrawalRequest, WithdrawalResponse, IncomingDepositRequest, IncomingDepositResponse, PlataYaDepositRequest, PlataYaDepositResponse, and TransferRequest to support updated transaction workflows.
- Updated ExternalBalanceValidationRequest and ExternalWalletValidationDTO to use Double instead of Float for amount and balance respectively.
- Removed the Transaction entity as part of the refactoring process.

+++ src/main/kotlin/com/plataya/external_api/model/dto/ExternalBalanceValidationRequest.kt
@@ -2,5 +2,5 @@ package com.plataya.external_api.model.dto
 
 data class ExternalBalanceValidationRequest(
     val cvu: Long,
-    val amount: Float
+    val amount: Double
 )
