+++ src/main/kotlin/com/plataya/external_api/model/dto/ExternalWalletValidationDTO.kt
@@ -3,6 +3,6 @@ package com.plataya.external_api.model.dto
 data class ExternalWalletValidationDTO(
     val cvu: Long,
     val exists: Boolean,
-    val balance: Float? = null,
+    val balance: Double? = null,
     val hasSufficientFunds: Boolean = false
 )
